Create Day11.py
Problem

Chef got his dream seat in F1 and secured a 3rd3^{rd}3rd place in his debut race. He now wants to know the time gap between him and the winner of the race.

You are his chief engineer and you only know the time gap between Chef and the runner up of the race, given as AAA seconds, and the time gap between the runner up and the winner of the race, given as BBB seconds.

Please calculate and inform Chef about the time gap between him and the winner of the race.
Input Format

    The first line of input will contain a single integer TTT, denoting the number of test cases.
    Each test case consists of a single line of input containing two space-separated integers AAA and BBB denoting the time gap between Chef and the runner up and the time gap between the runner up and the winner respectively.

Output Format

For each test case, output the time gap between Chef and the winner of the race.
Constraints

    1≤T≤1001 \leq T \leq 1001≤T≤100
    1≤A,B≤101 \leq A, B \leq 101≤A,B≤10

Sample 1:
Input
Output

4
1 1
2 5
3 2
5 6

2
7
5
11

Explanation:

Test case 111: The time gap between Chef and runner up is 111 second. The time gap between runner up and the winner is 111 second. Thus, the time gap between Chef and the winner is 1+1=21+1=21+1=2 seconds.

Test case 222: The time gap between Chef and runner up is 222 seconds. The time gap between runner up and the winner is 555 second. Thus, the time gap between Chef and the winner is 2+5=72+5=72+5=7 seconds.

Test case 333: The time gap between Chef and runner up is 333 seconds. The time gap between runner up and the winner is 222 second. Thus, the time gap between Chef and the winner is 3+2=53+2=53+2=5 seconds.

Test case 444: The time gap between Chef and runner up is 555 seconds. The time gap between runner up and the winner is 666 second. Thus, the time gap between Chef and the winner is 5+6=115+6=115+6=11 seconds.